// DXC_Frontend/src/api/models/getApiZaloMiniAppAdminFeedbackListParams.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */

export type GetApiZaloMiniAppAdminFeedbackListParams = {
  current?: number
  pageSize?: number
  title?: string
  fullName?: string
  phoneNumber?: string
  location?: string
  currentStatusId?: number
  assignedDepartmentPublicId?: string
  isPublic?: boolean
  isActive?: boolean
  createdFrom?: string
  createdTo?: string
}